Looked at effects of walker on signals that have periodic components but have dominating "noise" (e.g., environmental, instrumental error)
Made a script that looped through the variable star data. The data itself are the lightcurves of 50x50 pixel aperture, where each pixel has its own lightcurve. Pixels near the center (25,25) have a periodic component due to there being a variable star. Turns out  that at masses greater and ~20, the walker means of the center pixels are lower than the rest, with the lowest being pixel 25,25, which directly shows the variable star.

This is inconclusive however, as there are only around 5-10 pixels that display significant periodic motion. Furthermore, there are cases where some pixels have low w_means that are near the edges of the aperture. Need to research into this effect, by generating some time-series that fit this pattern.